TDL-12679: Added seperate integration test file for verify dynamic fields of organizations
